// types/projects.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Projects {
  projects: {
    [k: string]: Project;
  };
  [k: string]: unknown;
}
/**
 * This interface was referenced by `undefined`'s JSON-Schema definition
 * via the `patternProperty` "^.*$".
 */
export interface Project {
  name: string;
  logo?: string;
  description: string;
  type: string;
  categories: ("Privacy" | "Blockchain" | "Verification" | "Security" | "Learning" | "Other")[];
  applications: ("Finance" | "Health" | "Gov" | "Infra" | "Info" | "Other")[];
  tags: (
    | "Access Control"
    | "Adversarial"
    | "Anonymity"
    | "Attack"
    | "Benchmark"
    | "Binary"
    | "Bluetooth"
    | "Byzantine Resilience"
    | "Cloud"
    | "Code Analysis"
    | "Computer Vision"
    | "Concurrency"
    | "Consensus"
    | "CrowdSource"
    | "Data Structures"
    | "Database"
    | "Debug"
    | "Decentralized"
    | "Deep Neural Networks"
    | "Development"
    | "Distributed Learning"
    | "Encryption"
    | "Exploit"
    | "Failure"
    | "Features"
    | "Federated Learning"
    | "Fuzzing"
    | "Homomorphic Encryption"
    | "Image Classification"
    | "Images"
    | "Inductive Bias"
    | "Internet of Things"
    | "Locking"
    | "Low-Level"
    | "Machine Learning"
    | "Mobile App"
    | "Natural Language"
    | "Network"
    | "Optimization"
    | "Predictor"
    | "Protection"
    | "Protocol"
    | "PyTorch"
    | "Secure Multi-Party Computation"
    | "Server"
    | "Simulator"
    | "Static Analysis"
    | "Survey"
    | "TensorFlow"
    | "Tor"
    | "Zero-Knowledge Proofs"
  )[];
  date_added: string;
  layman_desc?: string;
  tech_desc?: string;
  notes?: string;
  url?: string;
  contacts?: {
    name: string;
    email?: string;
    url?: string;
    [k: string]: unknown;
  }[];
  code?: {
    type: string;
    url?: string;
    date_last_commit?: string;
    [k: string]: unknown;
  };
  doc?: string;
  lines_of_code?: number | string;
  c4dt_contact?: {
    name: string;
    email: string;
    [k: string]: unknown;
  };
  language?: string;
  license?: string;
  information?: {
    type: string;
    url: string;
    title: string;
    notes?: {
      label?: string;
      text: string;
      url?: string;
      [k: string]: unknown;
    }[];
    [k: string]: unknown;
  }[];
  maturity?: number;
  incubator?: {
    work: string;
    type?: "incubated" | "incubated_market" | "retired" | "retired_archived";
    products?: {
      type: "Demo" | "Hands-on" | "Pilot" | "App" | "Library" | "OSS-participation" | "Archive";
      url: string;
      title: string;
      code?: string;
      [k: string]: unknown;
    }[];
    [k: string]: unknown;
  };
  date_updated?: string;
  [k: string]: unknown;
}